Moves all PgQuery module to QueryBuilder
